Introduce FetchedDocument reusing part of Extractor interface
