+++ kifi-backend/modules/rover/app/com/keepit/rover/extractor/JsoupDocument.scala
@@ -8,28 +8,23 @@ import org.jsoup.nodes.Document
 import scala.collection.JavaConversions._
 import scala.util.Try
 
-class JsoupDocument(doc: Document) extends Logging {
+case class JsoupDocument(doc: Document) extends FetchedDocument with Logging {
 
   def getMetadata(name: String): Option[String] = {
-    if (name.toLowerCase == "title") {
-      toOption(doc.select("head title").text)
-    } else {
-      None
-    }
-      .orElse(toOption(doc.select("meta[name=" + name + "]").attr("content")))
-      .orElse(toOption(doc.select("meta[property=" + name + "]").attr("content")))
+    toOption(doc.select("meta[name=" + name + "]").attr("content")) orElse
+      toOption(doc.select("meta[property=" + name + "]").attr("content"))
   }
 
-  def getLinks(name: String): Set[String] = {
-    val urls = doc.select("link[ref=" + name + "]").iterator() map { e => e.attr("href") }
+  def getLinks(rel: String): Set[String] = {
+    val urls = doc.select("link[ref=" + rel + "]").iterator() map { e => e.attr("href") }
     urls filterNot { str => str == null || str.isEmpty } toSet
   }
+
+  def getTitle: Option[String] = toOption(doc.select("head title").text) orElse getMetadata("title")
 
   private def toOption(unsafe: => String): Option[String] = {
     Try(Option(unsafe)).toOption.flatten.filter(_.nonEmpty)
   }
-
-  def getKeywords(): Option[String] = getMetadata("keywords")
 }
 
 object JsoupDocument {
